Cors method allows (#9)
* Correction of testing preflight request declaring allowed methods

* example of logging config with spring boot
