+++ src/main/resources/application.properties
@@ -12,3 +12,6 @@ spring.jpa.properties.hibernate.dialect=@jdbc.dialect@
 
 # version applicative
 app.version=@app.version@
+
+# example logging configuration
+# logging.level.org.springframework.web.cors: DEBUG
